refactoring Form Sing In
