+++ app/src/main/java/ve/com/teeac/svgs/MainActivity.kt
@@ -3,11 +3,13 @@ package ve.com.teeac.svgs
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.compose.ui.ExperimentalComposeUiApi
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import ve.com.teeac.svgs.core.connection_network.presentation.ConnectivityStatus
 import ve.com.teeac.svgs.ui.theme.SVGSTheme
 
+@ExperimentalComposeUiApi
 @ExperimentalCoroutinesApi
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
