+++ app/src/main/java/ve/com/teeac/svgs/authentication/data/data_source/AuthResultContract.kt
@@ -1,4 +1,4 @@
-package ve.com.teeac.svgs.authentication.auth_google
+package ve.com.teeac.svgs.authentication.data.data_source
 
 import android.app.Activity
 import android.content.Context
