+++ app/src/main/java/ve/com/teeac/svgs/authentication/data/repository/AuthRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package ve.com.teeac.svgs.authentication.data.repository
 
 import kotlinx.coroutines.flow.SharedFlow
-import ve.com.teeac.svgs.authentication.auth_google.Credentials
 import ve.com.teeac.svgs.authentication.data.data_source.AuthRemoteUser
+import ve.com.teeac.svgs.authentication.data.data_source.Credentials
 import ve.com.teeac.svgs.authentication.data.models.UserInfo
 import ve.com.teeac.svgs.authentication.domain.repositories.AuthRepository
 import javax.inject.Inject
