+++ app/src/main/java/ve/com/teeac/svgs/authentication/domain/ValidationField.kt
@@ -1,4 +1,4 @@
-package ve.com.teeac.svgs.authentication
+package ve.com.teeac.svgs.authentication.domain
 
 import androidx.core.util.PatternsCompat
