+++ app/src/main/java/ve/com/teeac/svgs/authentication/domain/use_case/SignInTwitterUseCase.kt
@@ -1,7 +1,7 @@
-package ve.com.teeac.svgs.authentication.auth_twitter.domain
+package ve.com.teeac.svgs.authentication.domain.use_case
 
 import android.app.Activity
-import ve.com.teeac.svgs.authentication.auth_twitter.data.data_source.AuthenticationOAuthByFirebase
+import ve.com.teeac.svgs.authentication.data.data_source.AuthenticationOAuthByFirebase
 import ve.com.teeac.svgs.authentication.data.models.UserInfo
 import ve.com.teeac.svgs.di.AuthTwitter
 import javax.inject.Inject
