+++ app/src/main/java/ve/com/teeac/svgs/authentication/domain/use_case/SignInWithGoogleUseCase.kt
@@ -1,6 +1,6 @@
-package ve.com.teeac.svgs.authentication.auth_google.domain
+package ve.com.teeac.svgs.authentication.domain.use_case
 
-import ve.com.teeac.svgs.authentication.auth_google.Credentials
+import ve.com.teeac.svgs.authentication.data.data_source.Credentials
 import ve.com.teeac.svgs.authentication.data.models.UserInfo
 import ve.com.teeac.svgs.authentication.domain.repositories.AuthRepository
 import javax.inject.Inject
