+++ app/src/main/java/ve/com/teeac/svgs/authentication/presentation/google_button/GoogleAuthViewModel.kt
@@ -1,4 +1,4 @@
-package ve.com.teeac.svgs.authentication.auth_google
+package ve.com.teeac.svgs.authentication.presentation.google_button
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -6,7 +6,8 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import ve.com.teeac.svgs.authentication.auth_google.domain.SignInWithGoogleUseCase
+import ve.com.teeac.svgs.authentication.data.data_source.Credentials
+import ve.com.teeac.svgs.authentication.domain.use_case.SignInWithGoogleUseCase
 import ve.com.teeac.svgs.core.exceptions.ExceptionManager
 import javax.inject.Inject
