+++ app/src/main/java/ve/com/teeac/svgs/di/AppModule.kt
@@ -6,12 +6,12 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import ve.com.teeac.svgs.authentication.auth_twitter.data.data_source.AuthenticationOAuthByFirebase
-import ve.com.teeac.svgs.authentication.auth_twitter.domain.SignInTwitterUseCase
 import ve.com.teeac.svgs.authentication.data.data_source.AuthRemoteUser
+import ve.com.teeac.svgs.authentication.data.data_source.AuthenticationOAuthByFirebase
 import ve.com.teeac.svgs.authentication.data.repository.AuthRepositoryImpl
 import ve.com.teeac.svgs.authentication.domain.repositories.AuthRepository
 import ve.com.teeac.svgs.authentication.domain.use_case.SignInByEmailAndPasswordUseCase
+import ve.com.teeac.svgs.authentication.domain.use_case.SignInTwitterUseCase
 import ve.com.teeac.svgs.authentication.domain.use_case.SignUpByEmailAndPasswordUseCase
 import javax.inject.Qualifier
 import javax.inject.Singleton
