--- app/src/test/java/ve/com/teeac/svgs/ExampleUnitTest.kt
@@ -1,18 +1,0 @@
-package ve.com.teeac.svgs
-
-import org.junit.Assert.assertEquals
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-
-class ExampleUnitTest {
-
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
